handslab-trya-chat-agents/requirements.txt
langchain-core==0.3.15
langchain-aws==0.2.6
langgraph==0.2.45
boto3==1.35.36
python-dotenv==1.0.0
pydantic==2.9.2
redis==5.0.1
requests==2.31.0
